+++ src/portal/App.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Router, Switch, Route, RouteProps } from 'react-router';
 
-import '@portal/utils/event';
+import { postMessage } from '@portal/utils/event';
 import history from './history';
 import TLoadable from '@web/components/TLoadable';
 
@@ -17,24 +17,32 @@ const ActorEdit = TLoadable(() => import('@portal/actor/edit'));
 const GroupActorList = TLoadable(() => import('@portal/group/actor/list'));
 const GroupActorDetail = TLoadable(() => import('@portal/group/actor/detail'));
 const GroupActorEdit = TLoadable(() => import('@portal/group/actor/edit'));
+const NoteCreate = TLoadable(() => import('@portal/note/create'));
 
 interface TitleRouteProps extends RouteProps {
   title: string;
 }
-const TitleRoute = (route: TitleRouteProps) => {
+const TitleRoute: React.FC<TitleRouteProps> = React.memo((route) => {
+  const { path, title } = route;
+
   return (
     <Route
-      path={route.path}
+      path={path}
       render={(props) => {
-        document.title = route.title || 'TRPG Portal';
+        document.title = title || 'TRPG Portal';
+
+        postMessage('common::updatePath', {
+          title: document.title,
+          path,
+        });
 
         const Component = route.component;
-
         return <Component {...props}>{route.children}</Component>;
       }}
     />
   );
-};
+});
+TitleRoute.displayName = 'TitleRoute';
 
 class App extends React.Component {
   render() {
@@ -86,6 +94,11 @@ class App extends React.Component {
             title="团人物编辑"
             path="/group/:groupUUID/actor/:groupActorUUID/edit"
             component={GroupActorEdit}
+          />
+          <TitleRoute
+            title="创建笔记"
+            path="/note/create"
+            component={NoteCreate}
           />
         </Switch>
       </Router>